+++ astropy/visualization/wcsaxes/coordinate_helpers.py
@@ -261,7 +261,7 @@ class CoordinateHelper:
         else:
             raise TypeError("separator should be a string, a tuple, or None")
 
-    def set_format_unit(self, unit):
+    def set_format_unit(self, unit, decimal=None, show_decimal_unit=True):
         """
         Set the unit for the major tick labels.
 
@@ -269,8 +269,16 @@ class CoordinateHelper:
         ----------
         unit : class:`~astropy.units.Unit`
             The unit to which the tick labels should be converted to.
+        decimal : bool, optional
+            Whether to use decimal formatting. By default this is `False`
+            for degrees or hours (which therefore use sexagesimal formatting)
+            and `True` for all other units.
+        show_decimal_unit : bool, optional
+            Whether to include units when in decimal mode.
         """
         self._formatter_locator.format_unit = u.Unit(unit)
+        self._formatter_locator.decimal = decimal
+        self._formatter_locator.show_decimal_unit = show_decimal_unit
 
     def set_ticks(self, values=None, spacing=None, number=None, size=None,
                   width=None, color=None, alpha=None, exclude_overlapping=False):