+++ astropy/visualization/wcsaxes/tests/test_formatter_locator.py
@@ -122,19 +122,20 @@ class TestAngleFormatterLocator:
                                                     ('hh:mm:ss', '1h01m34s'),
                                                     ('hh:mm:ss.s', '1h01m34.1s'),
                                                     ('hh:mm:ss.ssss', '1h01m34.1354s'),
-                                                    ('d', '15'),
-                                                    ('d.d', '15.4'),
-                                                    ('d.dd', '15.39'),
-                                                    ('d.ddd', '15.392'),
-                                                    ('m', '924'),
-                                                    ('m.m', '923.5'),
-                                                    ('m.mm', '923.53'),
-                                                    ('s', '55412'),
-                                                    ('s.s', '55412.0'),
-                                                    ('s.ss', '55412.03'),
+                                                    ('d', '15\xb0'),
+                                                    ('d.d', '15.4\xb0'),
+                                                    ('d.dd', '15.39\xb0'),
+                                                    ('d.ddd', '15.392\xb0'),
+                                                    ('m', '924\''),
+                                                    ('m.m', '923.5\''),
+                                                    ('m.mm', '923.53\''),
+                                                    ('s', '55412"'),
+                                                    ('s.s', '55412.0"'),
+                                                    ('s.ss', '55412.03"'),
                                                     ])
     def test_format(self, format, string):
         fl = AngleFormatterLocator(number=5, format=format)
+        print(fl.formatter([15.392231] * u.degree, None, format='ascii')[0], string)
         assert fl.formatter([15.392231] * u.degree, None, format='ascii')[0] == string
 
     @pytest.mark.parametrize(('separator', 'format', 'string'), [(('deg', "'", '"'), 'dd', '15deg'),
@@ -144,8 +145,8 @@ class TestAngleFormatterLocator:
                                                                  (':', 'dd:mm:ss.s', '15:23:32.0'),
                                                                  ((':', ":", 's'), 'hh', '1:'),
                                                                  (('-', "-", 's'), 'hh:mm:ss.ssss', '1-01-34.1354s'),
-                                                                 (('d', ":", '"'), 'd', '15'),
-                                                                 (('d', ":", '"'), 'd.d', '15.4'),
+                                                                 (('d', ":", '"'), 'd', '15\xb0'),
+                                                                 (('d', ":", '"'), 'd.d', '15.4\xb0'),
                                                                  ])
     def test_separator(self, separator, format, string):
         fl = AngleFormatterLocator(number=5, format=format)
@@ -240,30 +241,44 @@ class TestAngleFormatterLocator:
         fl = AngleFormatterLocator()
         assert fl.formatter([15.392231] * u.degree, spacing)[0] == string
 
-    @pytest.mark.parametrize(('format_unit', 'decimal', 'spacing', 'string'),
-                             [(u.degree, False, 2 * u.degree, '15\xb0'),
-                              (u.degree, False, 2 * u.arcmin, '15\xb024\''),
-                              (u.degree, False, 2 * u.arcsec, '15\xb023\'32"'),
-                              (u.degree, False, 0.1 * u.arcsec, '15\xb023\'32.0"'),
-                              (u.hourangle, False, 15 * u.degree, '1h'),
-                              (u.hourangle, False, 15 * u.arcmin, '1h02m'),
-                              (u.hourangle, False, 15 * u.arcsec, '1h01m34s'),
-                              (u.hourangle, False, 1.5 * u.arcsec, '1h01m34.1s'),
-                              (u.degree, True, 15 * u.degree, '15'),
-                              (u.degree, True, 0.12 * u.degree, '15.39'),
-                              (u.degree, True, 0.0036 * u.arcsec, '15.392231'),
-                              (u.arcmin, True, 15 * u.degree, '924'),
-                              (u.arcmin, True, 0.12 * u.degree, '923.5'),
-                              (u.arcmin, True, 0.1 * u.arcmin, '923.5'),
-                              (u.arcmin, True, 0.0002 * u.arcmin, '923.5339'),
+    @pytest.mark.parametrize(('format_unit', 'decimal', 'show_decimal_unit', 'spacing', 'ascii', 'latex'),
+                             [(u.degree, False, True, 2 * u.degree, '15\xb0', '$15^\circ$'),
+                              (u.degree, False, True, 2 * u.arcmin, '15\xb024\'', '$15^\circ24{}^\prime$'),
+                              (u.degree, False, True, 2 * u.arcsec, '15\xb023\'32"', '$15^\circ23{}^\prime32{}^{\prime\prime}$'),
+                              (u.degree, False, True, 0.1 * u.arcsec, '15\xb023\'32.0"', '$15^\circ23{}^\prime32.0{}^{\prime\prime}$'),
+                              (u.hourangle, False, True, 15 * u.degree, '1h', '$1^\mathrm{h}$'),
+                              (u.hourangle, False, True, 15 * u.arcmin, '1h02m', '$1^\mathrm{h}02^\mathrm{m}$'),
+                              (u.hourangle, False, True, 15 * u.arcsec, '1h01m34s', '$1^\mathrm{h}01^\mathrm{m}34^\mathrm{s}$'),
+                              (u.hourangle, False, True, 1.5 * u.arcsec, '1h01m34.1s', '$1^\mathrm{h}01^\mathrm{m}34.1^\mathrm{s}$'),
+                              (u.degree, True, True, 15 * u.degree, '15\xb0', '$15\mathrm{^\circ}$'),
+                              (u.degree, True, True, 0.12 * u.degree, '15.39\xb0', '$15.39\mathrm{^\circ}$'),
+                              (u.degree, True, True, 0.0036 * u.arcsec, '15.392231\xb0', '$15.392231\mathrm{^\circ}$'),
+                              (u.arcmin, True, True, 15 * u.degree, '924\'', '$924\mathrm{^\prime}$'),
+                              (u.arcmin, True, True, 0.12 * u.degree, '923.5\'', '$923.5\mathrm{^\prime}$'),
+                              (u.arcmin, True, True, 0.1 * u.arcmin, '923.5\'', '$923.5\mathrm{^\prime}$'),
+                              (u.arcmin, True, True, 0.0002 * u.arcmin, '923.5339\'', '$923.5339\mathrm{^\prime}$'),
+                              (u.arcsec, True, True, 0.01 * u.arcsec, '55412.03"', '$55412.03\mathrm{^{\prime\prime}}$'),
+                              (u.arcsec, True, True, 0.001 * u.arcsec, '55412.032"', '$55412.032\mathrm{^{\prime\prime}}$'),
+                              (u.mas, True, True, 0.001 * u.arcsec, '55412032mas', '$55412032\mathrm{mas}$'),
+                              (u.degree, True, False, 15 * u.degree, '15', '15'),
+                              (u.degree, True, False, 0.12 * u.degree, '15.39', '15.39'),
+                              (u.degree, True, False, 0.0036 * u.arcsec, '15.392231', '15.392231'),
+                              (u.arcmin, True, False, 15 * u.degree, '924', '924'),
+                              (u.arcmin, True, False, 0.12 * u.degree, '923.5', '923.5'),
+                              (u.arcmin, True, False, 0.1 * u.arcmin, '923.5', '923.5'),
+                              (u.arcmin, True, False, 0.0002 * u.arcmin, '923.5339', '923.5339'),
+                              (u.arcsec, True, False, 0.01 * u.arcsec, '55412.03', '55412.03'),
+                              (u.arcsec, True, False, 0.001 * u.arcsec, '55412.032', '55412.032'),
+                              (u.mas, True, False, 0.001 * u.arcsec, '55412032', '55412032'),
                               # Make sure that specifying None defaults to
                               # decimal for non-degree or non-hour angles
-                              (u.arcsec, None, 0.01 * u.arcsec, '55412.03')])
-    def test_formatter_no_format_with_units(self, format_unit, decimal, spacing, string):
+                              (u.arcsec, None, True, 0.01 * u.arcsec, '55412.03"', '$55412.03\mathrm{^{\prime\prime}}$')])
+    def test_formatter_no_format_with_units(self, format_unit, decimal, show_decimal_unit, spacing, ascii, latex):
         # Check the formatter works when specifying the default units and
         # decimal behavior to use.
-        fl = AngleFormatterLocator(unit=u.degree, format_unit=format_unit, decimal=decimal)
-        assert fl.formatter([15.392231] * u.degree, spacing, format='ascii')[0] == string
+        fl = AngleFormatterLocator(unit=u.degree, format_unit=format_unit, decimal=decimal, show_decimal_unit=show_decimal_unit)
+        assert fl.formatter([15.392231] * u.degree, spacing, format='ascii')[0] == ascii
+        assert fl.formatter([15.392231] * u.degree, spacing, format='latex')[0] == latex
 
     def test_incompatible_unit_decimal(self):
         with pytest.raises(UnitsError) as exc: